Added processing and limited video recording for UIImagePickerViewController.
